+++ src/stdafx.h
@@ -15,20 +15,20 @@
 #include "IMemoryManager.h"
 
 #ifdef _WINDOWS
-    #include <windows.h>
+#include <windows.h>
 #else
-    typedef int64_t HWND;
+typedef int64_t HWND;
 #endif //_WINDOWS
 
 #if defined(__linux__) || defined(__APPLE__)
-	#define LINUX_OR_MACOS
+#define LINUX_OR_MACOS
 #endif
 
-std::wstring MB2WC(const std::string &source);
+std::wstring MB2WC(const std::string& source);
 
-std::string WC2MB(const std::wstring &source);
+std::string WC2MB(const std::wstring& source);
 
-long VarToInt(tVariant* paParams);
+int32_t VarToInt(tVariant* paParams);
 
 #ifdef _WINDOWS
 HWND VarToHwnd(tVariant* paParams);
